requirements.txt
langchain-openai
langchain
python-dotenv
langchain-community
langchain_core
